requirements.txt
# 識（しき）依存パッケージ

# Web Framework
fastapi==0.115.*
uvicorn[standard]==0.34.*

# LINE Bot SDK v3
line-bot-sdk==3.*

# LLM Providers
google-genai==1.*
openai>=1.0
anthropic>=0.40

# MCP（Model Context Protocol）
mcp>=1.0

# ブラウザ制御（Layer 2: Playwright headless）
playwright==1.*
playwright-stealth==2.*

# macOS操作（スクロール、マウス、画面サイズ取得）
# macOSの場合のみ必要。Windows/Linuxでは不要
pyobjc-framework-Quartz==12.*; sys_platform == "darwin"

# Windows/Linux操作（macOSでは不要）
pyautogui>=0.9; sys_platform != "darwin"
Pillow>=10.0; sys_platform != "darwin"
pyperclip>=1.8; sys_platform != "darwin"

# Discord Bot
discord.py>=2.3

# HTTP Client（Notion API等）
httpx>=0.27

# 環境変数
python-dotenv==1.*

# テスト
pytest>=8.0
pytest-asyncio>=0.23
